+++ src/Plugin.php
@@ -37,6 +37,7 @@ use craft\commerce\fieldlayoutelements\PurchasableSkuField;
 use craft\commerce\fieldlayoutelements\PurchasableStockField;
 use craft\commerce\fieldlayoutelements\PurchasableWeightField;
 use craft\commerce\fieldlayoutelements\UserAddressSettings;
+use craft\commerce\fieldlayoutelements\UserCommerceField;
 use craft\commerce\fieldlayoutelements\VariantsField as VariantsLayoutElement;
 use craft\commerce\fieldlayoutelements\VariantTitleField;
 use craft\commerce\fields\Products as ProductsField;
@@ -300,7 +301,6 @@ class Plugin extends BasePlugin
             $this->_registerWidgets();
             $this->_registerElementExports();
             $this->_defineFieldLayoutElements();
-            $this->_registerTemplateHooks();
             $this->_registerRedactorLinkOptions();
             $this->_registerCKEditorLinkOptions();
         } else {
@@ -1013,6 +1013,13 @@ class Plugin extends BasePlugin
                 case Address::class:
                     $e->fields[] = UserAddressSettings::class;
                     break;
+                case UserElement::class:
+                    // todo: remove in favor of a dedicated user management screen
+                    $currentUser = Craft::$app->getUser()->getIdentity();
+                    if ($currentUser?->can('commerce-manageOrders') || $currentUser?->can('commerce-manageSubscriptions')) {
+                        $e->fields[] = UserCommerceField::class;
+                    }
+                    break;
                 case Product::class:
                     $e->fields[] = ProductTitleField::class;
                     $e->fields[] = VariantsLayoutElement::class;
@@ -1083,18 +1090,5 @@ class Plugin extends BasePlugin
                 'helpSummary' => 'Re-saves Commerce carts.',
             ];
         });
-    }
-
-    /**
-     * Registers templates hooks for inserting Commerce information in the control panel
-     *
-     * @since 2.2
-     */
-    private function _registerTemplateHooks(): void
-    {
-        if ($this->getSettings()->showEditUserCommerceTab) {
-            Craft::$app->getView()->hook('cp.users.edit', [$this->getCustomers(), 'addEditUserCommerceTab']);
-            Craft::$app->getView()->hook('cp.users.edit.content', [$this->getCustomers(), 'addEditUserCommerceTabContent']);
-        }
     }
 }